+++ receiver/f12021/packets/helpers.py
@@ -2,9 +2,11 @@ from lib.logger import log
 from .base import PacketHeader
 from .session import PacketSessionData
 from .lap import PacketLapData
+from .event import PacketEventData
 from .participants import PacketParticipantsData
 from .setup import PacketCarSetupData
 from .telemetry import PacketCarTelemetryData
+from .car_status import PacketCarStatusData
 from .final_classification import PacketFinalClassificationData
 from .session_history import PacketSessionHistoryData
 
@@ -12,9 +14,11 @@ from .session_history import PacketSessionHistoryData
 HeaderFieldsToPacketType = {
      1: PacketSessionData,
      2: PacketLapData,
+     3: PacketEventData,
      4: PacketParticipantsData,
      5: PacketCarSetupData,
      6: PacketCarTelemetryData,
+     7: PacketCarStatusData,
      8: PacketFinalClassificationData,
     11: PacketSessionHistoryData
 }