+++ receiver/receiver.py
@@ -118,18 +118,23 @@ class RaceReceiver(threading.Thread):
         while not self.kill_event.is_set():
             try:
                 incoming_udp_packet = self.udp_socket.recv(2048)
-                if not self.processor:
-                    # Get game version -- raises if unknown or not found
-                    game_version = parse_game_version_from_udp_packet(incoming_udp_packet)
-                    if game_version == "f12020":
-                        log.info("Detected F1 2020 game version")
+                # Get game version -- raises if unknown or not found
+                # Do this for every packet so that we can handle game switches in flight
+                game_version = parse_game_version_from_udp_packet(incoming_udp_packet)
+                if game_version == "f12020":
+                    # Only start processor if it's not set yet or has switched
+                    if not self.processor or not isinstance(self.processor, F12020Processor):
+                        log.info("Detected F1 2020 game version, starting F1 2020 processor.")
                         self.processor = F12020Processor(self.f1laps_api_key, self.telemetry_enabled)
-                    elif game_version == "f12021":
-                        log.info("Detected F1 2021 game version")
+                elif game_version == "f12021":
+                    # Only start processor if it's not set yet or has switched
+                    if not self.processor or not isinstance(self.processor, F12021Processor):
+                        log.info("Detected F1 2021 game version, starting F1 2021 processor.")
                         self.processor = F12021Processor(self.f1laps_api_key, self.telemetry_enabled)
                 if self.processor:
                     self.processor.process(incoming_udp_packet)
             except Exception as ex:
+                log.info(ex)
                 sentry_sdk.capture_exception(ex)